+++ ReleaseNotes.md
@@ -10,6 +10,7 @@
 * Add `Stream.ReadAsync`/`WriteAsync` overloads for `Memory<byte>` on .NET Standard 2.1 and .NET Core.
 * Obsolete `DictionaryUtility.CreateKeyValuePair` on .NET Standard 2.1 and .NET Core.
 * Obsolete `EnumerableUtility.EnumerateBatches` on .NET 6.
+* Obsolete `EnumerableUtility.DistinctBy` on .NET 6.
 
 ## 0.11.2